### 2017-12-18 #173 @chris-prener
driftR Pre-Submission
This is a *pre-submission* inquiry based on a Twitter exchange with @maelle - we are interested to hear your thoughts on whether this package fits with ropensci!

A tidy implementation of equations that correct for instrumental drift in continuous water quality monitoring data. There are many sources of water quality data, each of which are susceptible to errors/inaccuracies due to drift. This package allows the user to correct their data using one or two standard reference values in a uniform, reproducible way.

```
Package: driftR
Title: Drift Correcting Water Quality Data
Version: 1.0.0
    person("Andrew", "Shaughnessy", , "andrew.shaughnessy@slu.edu", c("aut", "cre")),
    person("Christopher", "Prener", , "chris.prener@slu.edu", c("aut")),
    person("Elizabeth", "Hasenmueller", , "elizabeth.hasenmueller@slu.edu", c("aut"))
    )
Description: A tidy implementation of equations that correct for instrumental drift in
    continuous water quality monitoring data.  There are many sources of water quality data
    including private (ex: YSI instruments) and open source (ex: USGS and NDBC), each of
    which are susceptible to errors/inaccuracies due to drift. This package allows the
    user to correct their data using one or two standard reference values in a uniform,
    reproducible way. The equations implemented are from Hasenmueller (2011)
    <doi:10.7936/K7N014KS>.
Imports: dplyr,
    glue
Suggests: testthat,
    covr
```

https://github.com/shaughnessyar/driftR

**Data Extraction**: The package includes a function for parsing output from a particular type of water quality monitoring instrument. We are planning to add additional functions to parse related instruments from different models / brands.

The target audience are hydrology and geochemistry researchers who use continuous water quality monitoring instruments.

There are no R packages that accomplish the same thing. There are some very expensive proprietary packages for applying correction functions, and USGS has its own proprietary system that cannot be distributed publicly. Thus `driftR` fills an important niche for hydrologists and geoscientists.

This is a *pre-submission* inquiry based on a Twitter exchange with @maelle

- [x] Do you intend for this package to go on CRAN?  - **Already on CRAN**
    - [ ] The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with a high-level description in the package root or in `inst/`. - **We will do this based on the outcome of the presubmission inquiry**
	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

Our package name does contain an upper-case letter, and our function names do combine snake case with camelCase - e.g. `dr_correctOne()`
