### 2018-01-10 #173 @maelle
@chris-prener thanks for considering a submission! Yes it'd fit in the data munging category but will be a more interesting submission once it handles several instruments which is your plan anyway. So submit it once it has a few more parsing functions for other instruments, so that the reviewed version might already be close to your "dream version". 😺
