### 2018-01-10 #173 @maelle
If you have any question in the meantime you can ping me here or ask in https://discuss.ropensci.org
